Fix timing bug in CPU logic
Now, after a second rework of the main CPU logic, we now pass
Blargg's cpu_instrs AND the instr_timing test! This is a success!

We pass all but three of the Mooneye timer tests, so the next step
is to work thorugh that to ensure that all parts implented so
far are correct.